Изменены названия функций для alias api и hashtable api. Изменено возвращаемое значение hash_function на unsigned_int Добавлены проверки при free в insert_hash_date Удалена бесполезная печать gaypromt в print_hash_table
