+++ src/exec/builtin_alias.c
@@ -6,7 +6,7 @@
 /*   By: nriker <nriker@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/28 11:34:50 by anton             #+#    #+#             */
-/*   Updated: 2021/01/07 00:04:33 by nriker           ###   ########.fr       */
+/*   Updated: 2021/01/07 10:31:20 by nriker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,21 +14,17 @@
 
 int			builtin_alias(char **args, t_env env, int subshell)
 {
-	t_hashalias		*alias;
-
-	if ((alias = static_hashalias_action(get)) == NULL)
-		return (EXIT_FAILURE);
 	(void)env;
 	(void)subshell;
 	if (*args == NULL)
-		alias_api_print_all_aliases(alias);
+		print_all_aliases();
 	else
 		while (*args)
 		{
 			if (*args && (args[1] == NULL || ft_strcmp(args[1], "=")))
-				alias_api_print_alias(*args);
+				print_alias(*args);
 			else if (*args && !ft_strcmp(args[1], "="))
-				alias_api_insert_alias(alias, *args, args[2]);
+				insert_alias(*args, args[2]);
 			args++;
 		}
 	return (EXIT_FAILURE);
