+++ src/exec/delete_alias.c
@@ -1,20 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   unalias_api_delete_alias.c                         :+:      :+:    :+:   */
+/*   delete_alias.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nriker <nriker@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/05 13:26:45 by nriker            #+#    #+#             */
-/*   Updated: 2021/01/05 13:32:13 by nriker           ###   ########.fr       */
+/*   Updated: 2021/01/07 11:09:32 by nriker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "t_hashalias.h"
 
-int						unalias_api_delete_alias(t_hashalias **hash_alias, char *key)
+int						delete_alias(char *key)
 {
-	if ((hash_api_delete_hash_table(key, &((*hash_alias)->hd))) == EXIT_FAILURE)
+	t_hashalias		*hash_alias;
+
+	if ((hash_alias = static_hashalias_action(get)) == NULL)
+		return (EXIT_FAILURE);
+	if ((delete_t_hashtable(key, &(hash_alias->hd))) == EXIT_FAILURE)
 		return (EXIT_FAILURE);
 	return (EXIT_SUCCESS);
 }
