+++ src/exec/delete_t_hashalias.c
@@ -1,26 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   alias_api_insert_alias.c                           :+:      :+:    :+:   */
+/*   delete_t_hashalias.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nriker <nriker@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/01/05 12:24:48 by nriker            #+#    #+#             */
-/*   Updated: 2021/01/05 12:30:37 by nriker           ###   ########.fr       */
+/*   Created: 2021/01/05 13:42:28 by nriker            #+#    #+#             */
+/*   Updated: 2021/01/07 11:08:05 by nriker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "t_hashalias.h"
 
-void					alias_api_insert_alias(t_hashalias *hash_alias,
-							char *key, char *value)
+void					delete_t_hashalias(void)
 {
-	if (value == NULL)
-	{
-		value = ft_strdup("''");
-		hash_api_insert_hash_value(key, value, hash_alias->hd);
-		free(value);
-	}
-	else
-		hash_api_insert_hash_value(key, value, hash_alias->hd);
+    t_hashalias		*hash_alias;
+
+	if ((hash_alias = static_hashalias_action(get)) == NULL)
+		return ;
+    if (hash_alias->hd != NULL)
+        delete_t_hashdata((&(hash_alias->hd)));
+    free(hash_alias);
 }
