+++ src/exec/search_alias.c
@@ -1,20 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   alias_api_delete_hashalias.c                       :+:      :+:    :+:   */
+/*   search_alias.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nriker <nriker@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/01/05 13:42:28 by nriker            #+#    #+#             */
-/*   Updated: 2021/01/05 13:45:55 by nriker           ###   ########.fr       */
+/*   Created: 2021/01/05 10:59:47 by nriker            #+#    #+#             */
+/*   Updated: 2021/01/07 11:03:52 by nriker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "t_hashalias.h"
 
-void					alias_api_delete_hashalias(t_hashalias **hash_alias)
+char					*search_alias(char *key)
 {
-    if ((*hash_alias)->hd != NULL)
-        hash_api_delete_hash_date((&(*hash_alias)->hd));
-    free(*hash_alias);
+	t_hashalias		*hash_alias;
+
+	if ((hash_alias = static_hashalias_action(get)) == NULL)
+		return (NULL);
+	return (search_hash_data(key, hash_alias->hd));
 }
