+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: nriker <nriker@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/24 20:16:51 by anton             #+#    #+#             */
-/*   Updated: 2021/01/05 11:49:30 by nriker           ###   ########.fr       */
+/*   Updated: 2021/01/07 10:56:48 by nriker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ int		main(int ac, char **av, char **environ)
 		init_terminal();
 		init_prompt();
 	}
-	// set_signal();
+	set_signal();
 	repl(tty_input);
 	return (0);
 }
